chore: move type guards to lib/learn-client/types
